+++ but/let-my-vue-components-look-stupid.js
@@ -1,0 +1,54 @@
+"use strict";
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+module.exports =
+{
+	"plugins" :
+	[
+		"vue"
+	],
+
+	"rules" :
+	{
+		"vue/attribute-hyphenation"                     : "off",
+		"vue/html-closing-bracket-newline"              : "off",
+		"vue/html-closing-bracket-spacing"              : "off",
+		"vue/html-indent"                               : "off",
+		"vue/html-quotes"                               : "off",
+		"vue/html-self-closing"                         : "off",
+		"vue/max-attributes-per-line"                   : "off",
+		"vue/multiline-html-element-content-newline"    : "off",
+		"vue/mustache-interpolation-spacing"            : "off",
+		"vue/name-property-casing"                      : "off",
+		"vue/no-multi-spaces"                           : "off",
+		"vue/no-spaces-around-equal-signs-in-attribute" : "off",
+		"vue/no-template-shadow"                        : "off",
+		"vue/prop-name-casing"                          : "off",
+		"vue/v-bind-style"                              : "off",
+		"vue/v-on-style"                                : "off",
+		"vue/attributes-order"                          : "off",
+		"vue/order-in-components"                       : "off",
+		"vue/array-bracket-spacing"                     : "off",
+		"vue/arrow-spacing"                             : "off",
+		"vue/block-spacing"                             : "off",
+		"vue/brace-style"                               : "off",
+		"vue/camelcase"                                 : "off",
+		"vue/comma-dangle"                              : "off",
+		"vue/component-name-in-template-casing"         : "off",
+		"vue/key-spacing"                               : "off",
+		"vue/keyword-spacing"                           : "off",
+		"vue/object-curly-spacing"                      : "off",
+		"vue/require-direct-export"                     : "off",
+		"vue/script-indent"                             : "off",
+		"vue/space-infix-ops"                           : "off",
+		"vue/space-unary-ops"                           : "off",
+		"vue/v-slot-style"                              : "off",
+		"vue/component-definition-name-casing"          : "off",
+		"vue/component-tags-order"                      : "off",
+		"vue/max-len"                                   : "off",
+		"vue/no-static-inline-styles"                   : "off",
+		"vue/require-name-property"                     : "off",
+		"vue/static-class-names-order"                  : "off"
+	}
+};